requirements.txt
pandas==1.5.3
nltk==3.8.1
scikit-learn==1.2.2
spacy==3.5.2
sentence-transformers==2.2.2
transformers==4.30.2
tqdm==4.65.0
Pillow==9.5.0
numpy==1.21.4
#en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
